feat(app, home): главная перенесена из app в homepage, настроен роутинг
