+++ package.json
@@ -31,6 +31,7 @@
 		"react-dnd-html5-backend": "^16.0.1",
 		"react-dom": "^18.2.0",
 		"react-redux": "^9.2.0",
+		"react-router": "^7.6.2",
 		"redux": "^5.0.1"
 	},
 	"devDependencies": {
