+++ src/pages/index.js
@@ -1,0 +1,3 @@
+import { HomePage } from './home';
+
+export { HomePage };
